Add support for specifying postgres schema
When working with database schema other than `public` (the default), it
is desirable to specify which schema to target for cleaning tables. This
adds support for passing a `schema` property in the `postgresql` config
argument of the `DatabaseCleaner` constructor.

Example:

```javascript
var config = {
  postgresql: {
    strategy: 'truncation',
    skipTables: [],
    schema: 'other_schema'
  }
};

databaseCleaner = new DatabaseCleaner('postgresql', config);
```

When specifying a `schema`, _only_ the specified `schema` will be
cleaned. When _not_ specifying a schema, _only_ the `public` schema will
be cleaned.